chore: add more test-suites
