--- packages/ui/components/menu/src/utils/uuid.js
@@ -1,3 +1,0 @@
-export function uuid() {
-  return Math.random().toString(36).substr(2, 10);
-}
