+++ packages/ui/components/menu/test/InteractiveListMixin.test.js
@@ -1,0 +1,3 @@
+import { runInteractiveListMixinSuite } from '../test-suites/InteractiveListMixin.suite.js';
+
+runInteractiveListMixinSuite();
