+++ packages/ui/components/menu/test/MoreButtonMenuMixin.test.js
@@ -1,0 +1,4 @@
+import { runMoreButtonMenuMixinSuite } from '../test-suites/MoreButtonMenuMixin.suite.js';
+import '@lion/ui/define/lion-menu.js';
+
+runMoreButtonMenuMixinSuite({ tagString: 'lion-menu' });
