+++ packages/ui/components/menu/test/MultiLevelListMixin.test.js
@@ -1,0 +1,3 @@
+import { runMultiLevelListMixinSuite } from '../test-suites/MultiLevelListMixin.suite.js';
+
+runMultiLevelListMixinSuite();
