+++ packages/ui/components/menu/test/lion-menu-hybrid.test.js
@@ -1,0 +1,3 @@
+import { runLionMenuHybridSuite } from '../test-suites/LionMenuHybrid.suite.js';
+
+runLionMenuHybridSuite();
